moves setup to separate file, install coverage
